Excise explicit proto conversion from disk buffer, reorg
This commit removes explicit proto conversion from the disk buffer code, relying
instead on newly introduce TryFrom/TryInto definitions. The goal is to make the
buffers generic. That's now more possible.

I have also reorganized the leveldb code to make the compiler error messages I'm
expecting easier to cope with.
